+++ UVtools.WPF/Windows/MessageWindow.axaml.cs
@@ -1,13 +1,16 @@
+using System;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Layout;
 using Avalonia.Markup.Xaml;
+using UVtools.Core.SystemOS;
 using UVtools.WPF.Controls;
 
 namespace UVtools.WPF.Windows
 {
     public partial class MessageWindow : WindowEx
     {
+        #region Members
         private string? _headerIcon;
         private ushort _headerIconSize = 64;
         private string _headerTitle;
@@ -15,6 +18,13 @@ namespace UVtools.WPF.Windows
         private string _message;
 
         private readonly StackPanel _buttonsRightPanel;
+        #endregion
+
+        #region Properties
+        /// <summary>
+        /// Gets the pressed button
+        /// </summary>
+        public ButtonWithIcon? PressedButton { get; private set; }
 
         public string? HeaderIcon
         {
@@ -55,7 +65,9 @@ namespace UVtools.WPF.Windows
             get => _aboutButtonIsVisible;
             set => RaiseAndSetIfChanged(ref _aboutButtonIsVisible, value);
         }
+        #endregion
 
+        #region Constructor
         public MessageWindow()
         {
             InitializeComponent();
@@ -66,22 +78,23 @@ namespace UVtools.WPF.Windows
             DataContext = this;
         }
 
-        public MessageWindow(string title, string? headerIcon, string? headerTitle, string message, ButtonWithIcon[]? buttons = null) : this()
+        public MessageWindow(string title, string? headerIcon, string? headerTitle, string message, ButtonWithIcon[]? rightButtons = null) : this()
         {
             Title = title;
             HeaderIcon = headerIcon;
             HeaderTitle = headerTitle;
             Message = message;
 
-            if (buttons is not null && buttons.Length > 0)
+            if (rightButtons is not null && rightButtons.Length > 0)
             {
                 _buttonsRightPanel.Children.Clear();
-                _buttonsRightPanel.Children.AddRange(buttons);
+                _buttonsRightPanel.Children.AddRange(rightButtons);
 
-                foreach (var button in buttons)
+                foreach (var button in rightButtons)
                 {
                     button.Click += (sender, e) =>
                     {
+                        PressedButton = button;
                         if (e.Handled) return;
                         Close(button);
                         e.Handled = true;
@@ -91,32 +104,72 @@ namespace UVtools.WPF.Windows
         }
 
         public MessageWindow(string title, string message, ButtonWithIcon[]? buttons = null) : this(title, null, null, message, buttons) { }
-
-        /*protected override void OnOpened(EventArgs e)
-        {
-            base.OnOpened(e);
-        }*/
+        
 
         private void InitializeComponent()
         {
             AvaloniaXamlLoader.Load(this);
         }
+        #endregion
 
+        #region Methods
         public async void OpenAboutWindow()
         {
             await new AboutWindow().ShowDialog(this);
         }
+        #endregion
 
-        public static ButtonWithIcon CreateButton(string? text, string? icon, int padding = 10) => 
-            new()
-            {
-                Icon = icon, 
-                Text = text, 
-                VerticalAlignment = VerticalAlignment.Center,
-                Padding = new Thickness(padding)
-            };
+        #region Static methods
+        public static ButtonWithIcon CreateButtonFunc(string? text, string? icon, Func<bool> customAction, int padding = 10)
+        {
+            var button = CreateButton(text, icon, padding);
+            button.Click += (sender, e) => e.Handled = customAction.Invoke();
+            return button;
+        }
+
+        public static ButtonWithIcon CreateButtonAction(string? text, string? icon, Action customAction, int padding = 10)
+        {
+            var button = CreateButton(text, icon, padding);
+            button.Click += (sender, e) => customAction.Invoke();
+            return button;
+        }
+
+        public static ButtonWithIcon CreateButtonFunc(string? text, Func<bool> customAction, int padding = 10) => CreateButtonFunc(text, null, customAction, padding);
+        public static ButtonWithIcon CreateButtonAction(string? text, Action customAction, int padding = 10) => CreateButtonAction(text, null, customAction, padding);
+
+        public static ButtonWithIcon CreateButton(string? text, string? icon, int padding = 10) => new()
+        {
+            Icon = icon,
+            Text = text,
+            VerticalAlignment = VerticalAlignment.Center,
+            Padding = new Thickness(padding)
+        };
 
         public static ButtonWithIcon CreateButton(string? text, int padding = 10) => CreateButton(text, null, padding);
+
+
+        public static ButtonWithIcon CreateLinkButtonAction(string? text, string? icon, string url, Action customAction, int padding = 10)
+        {
+            var button = CreateButtonFunc(text, icon, () =>
+            {
+                customAction.Invoke();
+                SystemAware.OpenBrowser(url);
+                return true;
+            }, padding);
+            return button;
+        }
+
+        public static ButtonWithIcon CreateLinkButton(string? text, string? icon, string url, int padding = 10)
+        {
+            var button = CreateButtonFunc(text, icon, () =>
+            {
+                SystemAware.OpenBrowser(url);
+                return true;
+            }, padding);
+            return button;
+        }
+        public static ButtonWithIcon CreateLinkButtonAction(string? text, string url, Action customAction, int padding = 10) => CreateLinkButtonAction(text, null, url, customAction, padding);
+        public static ButtonWithIcon CreateLinkButton(string? text, string url, int padding = 10) => CreateLinkButton(text, null, url, padding);
 
         public static ButtonWithIcon CreateCancelButton(string? icon = null, int padding = 10)
         {
@@ -131,5 +184,6 @@ namespace UVtools.WPF.Windows
             btn.IsCancel = true;
             return btn;
         }
+        #endregion
     }
 }